Project Attempt 1/requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
backoff==2.2.1
bcrypt==4.0.1
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
choreographer==1.2.1
chromadb==0.3.23
click==8.3.1
clickhouse-connect==0.10.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
distro==1.9.0
dnspython==2.8.0
duckdb==1.4.2
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.128.0
filelock==3.20.0
flasgger==0.9.7.1
Flask==3.1.2
flask-sock==0.7.0
fonttools==4.60.1
fsspec==2025.10.0
greenlet==3.2.4
h11==0.16.0
hnswlib==0.8.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaleido==1.2.0
kiwisolver==1.4.9
logistro==2.0.1
lz4==4.4.5
MarkupSafe==3.0.3
matplotlib==3.10.7
minsearch @ git+https://github.com/alexeygrigorev/minsearch.git@cc48fb3e25c1b777f780779cd294a67ab8e327f3
mistune==3.1.4
MouseInfo==0.1.3
mpmath==1.3.0
narwhals==2.12.0
networkx==3.6
numpy==2.3.5
openai==2.14.0
openpyxl==3.1.5
orjson==3.11.4
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==12.0.0
plotly==6.5.0
pluggy==1.6.0
posthog==7.0.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
PyAutoGUI==0.9.54
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
PyMySQL==1.1.2
pyparsing==3.2.5
pypdf==6.5.0
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==9.0.1
pytest-asyncio==1.3.0
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.29.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
sentence-transformers==5.1.2
setuptools==80.9.0
simple-websocket==1.1.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.4
starlette==0.50.0
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uv==0.9.13
uvicorn==0.40.0
vanna==0.7.9
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.3.2
zstandard==0.25.0
